refactor: separating components
