+++ src/app/login/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+const LoginPage = () => {
+  return <div>Welcome to the Login page</div>;
+};
+
+export default LoginPage;
